pyproject.toml
[project]
name = "redis-nbp"
version = "0.1.0"
description = "Food Recipe RAG Chat — FastAPI + LangGraph + Redis"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "streamlit>=1.38.0",
    "redis[hiredis]>=5.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-redis>=0.2.0",
    "langgraph>=0.2.0",
    "pydantic-settings>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.26.0",
]

[tool.uv]
package = false
